requirements.txt
aiohttp
aiosqlite
discord.py
python-dotenv
streamrip>=2.0
